backend/requirements.txt
alembic==1.13.1
anyio==4.2.0
asgiref==3.7.2
Authlib==1.2.1
azure-common==1.1.28
azure-core==1.29.6
azure-identity==1.15.0
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.0.1
azure-mgmt-subscription==3.1.1
Babel==2.14.0
bcrypt==4.0.1
bidict==0.22.1
blinker==1.7.0
boto3==1.33.13
botocore==1.33.13
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
Django==5.0
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.4.2
email-validator==2.1.0.post1
eventlet==0.33.3
exceptiongroup==1.2.0
Flask==2.3.3
flask-babel==4.0.0
Flask-Compress==1.14
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.2.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.62.0
greenlet==1.1.2
h11==0.14.0
httpagentparser==1.9.5
httplib2==0.22.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.3.0
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
keyring==24.3.0
ldap3==2.9.1
Mako==1.3.0
MarkupSafe==2.1.3
more-itertools==10.1.0
msal==1.26.0
msal-extensions==1.1.0
msrest==0.7.1
oauthlib==3.2.2
packaging==23.2
paramiko==3.4.0
passlib==1.7.4
pgadmin4==8.1
portalocker==2.8.2
protobuf==4.25.1
psutil==5.9.7
psycopg==3.1.12
psycopg-binary==3.1.12
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==3.1.1
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.1
python-socketio==5.10.0
pytz==2023.3.post1
qrcode==7.4.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.8.2
SecretStorage==3.3.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
speaklater3==1.4
SQLAlchemy==2.0.24
sqlparse==0.4.4
sshtunnel==0.4.0
typing_extensions==4.9.0
ua-parser==0.18.0
uritemplate==4.1.1
urllib3==1.26.18
user-agents==2.2.0
Werkzeug==2.3.8
wsproto==1.2.0
WTForms==3.1.1
zipp==3.17.0
